# src/app/docs/swagger-suno-api.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "suno-api",
    "description": "Use API to call the music generation service of Suno.ai and easily integrate it into agents like GPTs.",
    "version": "1.1.0"
  },
...
  "paths": {
...
    "/api/generate_sound": {
      "post": {
        "summary": "Generate sound effects based on Prompt.",
        "description": "Generate short sound effects, loops, and one-shots based on a text description.\n\n2 audio files will be generated for each request, consuming a total of 10 credits.\n\n`wait_audio` can be set to API mode:\n\n• By default, it is set to `false`, which indicates the background mode. It will only return audio task information, and you will need to call the get API to retrieve detailed audio information.\n\n• If set to `true`, it simulates synchronous mode. The API will wait for a maximum of 100s until the audio is generated, and will directly return the audio link and other information. Recommend using in GPTs and other agents.",
        "tags": ["default"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["prompt"],
                "properties": {
                  "prompt": {
                    "type": "string",
                    "description": "Prompt describing the sound effect",
                    "example": "A futuristic laser gun shooting sound effect"
                  },
                  "loop": {
                    "type": "boolean",
                    "description": "Whether to generate a looping sound effect",
                    "example": false
                  },
                  "model": {
                    "type": "string",
                    "description": "Model name, default is chirp-hawk",
                    "example": "chirp-hawk"
                  },
                  "tempo": {
                    "type": "number",
                    "description": "BPM of the generated sound effect, integer 1-300. Omit for auto.",
                    "example": 120
                  },
                  "key": {
                    "type": "string",
                    "description": "Musical key of the generated sound effect, e.g. 'C', 'F#' or 'A#m' (m = minor). Omit for any key.",
                    "example": "C"
                  },
                  "wait_audio": {
                    "type": "boolean",
                    "description": "Whether to wait for audio generation, default is false, directly return audio task information; set to true, will wait for up to 100s until the audio is generated.",
                    "example": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": ["0", "1"],
                    "properties": [
                      {
                        "$ref": "#/components/schemas/audio_info"
                      },
                      {
                        "$ref": "#/components/schemas/audio_info"
                      }
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "audio_info": {
        "type": "object",
        "required": [
          "id",
          "title",
          "image_url",
          "lyric",
          "audio_url",
          "video_url",
          "created_at",
          "model_name",
          "status",
          "gpt_description_prompt",
          "prompt",
          "type",
          "tags"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "audio id"
          },
          "title": {
            "type": "string",
            "description": "music title"
          },
          "image_url": {
            "type": "string",
            "description": "music cover image"
          },
          "lyric": {
            "type": "string",
            "description": "music lyric"
          },
          "audio_url": {
            "type": "string",
            "description": "music download url"
          },
          "video_url": {
            "type": "string",
            "description": "Music video download link, can be used to share"
          },
          "created_at": {
            "type": "string",
            "description": "Create time"
          },
          "model_name": {
            "type": "string",
            "description": "suno model name, chirp-v3"
          },
          "status": {
            "type": "string",
            "description": "The generated states include submitted, queue, streaming, complete."
          },
          "gpt_description_prompt": {
            "type": "string",
            "description": "Simple mode on user input prompt, Suno will generate formal prompts, lyrics, etc."
          },
          "prompt": {
            "type": "string",
            "description": "The final prompt for executing the generation task, customized by the user in custom mode, automatically generated by Suno in simple mode."
          },
          "type": {
            "type": "string",
            "description": "Type"
          },
          "tags": {
            "type": "string",
            "description": "Music genre. User-provided in custom mode, automatically generated by Suno in simple mode."
          }
        },
        "title": "audio_info",
        "description": "Audio Info"
      }
    }
  }
}